adding connection_cache resolver. moved connections inside master_api so scripts can have access to them directly. now gateway discover connection_cache topics. no need to remap -> dropping roslaunch args.
